requirements.txt
absl-py==0.10.0
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
batchgenerators==0.21
bleach==3.2.1
blinker==1.4
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
compress-json==1.0.4
configparser==5.0.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dicom-numpy==0.4.0
entrypoints==0.3
future==0.18.2
gast==0.3.3
google-auth==1.21.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
importlib-metadata==1.7.0
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.1
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
linecache2==1.0.0
llvmlite==0.34.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.2
mistune==0.8.4
nbclient==0.5.0
nbconvert==6.0.5
nbformat==5.0.7
nest-asyncio==1.4.0
networkx==2.5
nibabel==3.1.1
notebook==6.1.4
numba==0.51.2
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.44
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.2
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.7
protobuf==3.13.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydicom==2.0.0
pydot==1.4.1
Pygments==2.7.1
pynrrd==0.4.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
sanitize-ml-labels==1.0.12
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.4.1
Send2Trash==1.5.0
SimpleITK==1.2.4
six==1.15.0
sklearn==0.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.9.3
tornado==6.0.4
tqdm==4.50.0
traceback2==1.4.0
traitlets==4.3.3
transformations==2020.1.1
unittest2==1.1.0
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.1.0
